Refactor: Regex 관련 함수명 수정
